gae_lycaptcha/requirements.txt
torch
torchvision
matplotlib
Flask==2.0.2
google-cloud-logging